+++ almost_make/utils/shellUtil/shellUtil.py
@@ -105,10 +105,10 @@ CUSTOM_COMMANDS = \
 }
 
 def customExit(args, stdin, stdout, stderr, state):
-    if len(args) == 0:
+    if len(args) == 1:
         sys.exit(0)
     else:
-        sys.exit(int(args[0]))
+        return int(args[1])
 
 LS_DIRECTORY_COLOR = FORMAT_COLORS['BLUE']
 LS_LINK_COLOR = FORMAT_COLORS['BLUE']
@@ -246,6 +246,9 @@ def customCd(args, stdin, stdout, stderr, state):
 
 def customEcho(args, stdin, stdout, stderr, state):
     if len(args) == 1:
+        # No arguments? Print a newline.
+        cprint('\n', file=stdout)
+
         return 0
     
     doEscapes = False
@@ -392,6 +395,10 @@ def customGrep(args, stdin, stdout, stderr, state):
                 flags = flags | re.IGNORECASE
 
             patterns.append(re.compile(part, flags))
+    else:
+        # If no default arguments, grep was probably called
+        # with an empty pattern. Note this.
+        patterns.append(re.compile('')) 
     
     def matchesLine(line):
         matches = []
@@ -585,7 +592,7 @@ def getCustomCommands(macros):
         addCustomCommand("ls", 1, customLs)
         result["dir"] = result["ls"]
         addCustomCommand("pwd", 1, customPwd)
-        addCustomCommand("echo", 2, customEcho)
+        addCustomCommand("echo", 1, customEcho)
         addCustomCommand("touch", 2, customTouch)
         addCustomCommand("cat", 2, customCat)
         addCustomCommand("grep", 2, customGrep)